fix score calc

+++ checks/sast.go
@@ -62,8 +62,7 @@ func SAST(c *checker.CheckRequest) checker.CheckResult {
 		// SAST run on commits.
 		result = r2
 
-		//nolint
-		result.Score = r2.Score/10*5 + r1.Score/10*5
+		result.Score = int(float32(r2.Score)/10*5 + float32(r1.Score)/10*5)
 		result.Reason = "not all commits are checked with a SAST tool"
 		i := strings.Index(r1.Reason, "-- score normalized")
 		if i < 0 {
